Update alarms perspective after blockserver update
