+++ base/uk.ac.stfc.isis.ibex.alarm/src/uk/ac/stfc/isis/ibex/alarm/AlarmCounter.java
@@ -102,8 +102,6 @@ public class AlarmCounter extends ModelObject {
                 firePropertyChange("alarmCount", prevCount, newCount);
             }
 		});
-
-        AlarmViewReloadManager.queueDelayedUpdate();
 	}
 
 	/**
@@ -112,33 +110,5 @@ public class AlarmCounter extends ModelObject {
 	public int getCount() {
 		return count;
 	}
-
-    private static class AlarmViewReloadManager {
-        private static Thread thread;
-
-        private static void queueDelayedUpdate() {
-            if (thread != null && thread.isAlive()) {
-                thread.interrupt();
-            }
-
-            thread = new Thread(reloadRunnable);
-
-            thread.start();
-        }
-
-        private static final Runnable reloadRunnable = new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    Thread.sleep(1000 * 30);
-                } catch (InterruptedException e) {
-                    return;
-                }
-                Alarm.getInstance().reload();
-            }
-        };
-
-    }
-
 
 }
